Refactor modules into utils/ and add documentation
