+++ server/api.js
@@ -1,6 +1,6 @@
 import { Router } from "express";
 
-import logger from "./logger";
+import logger from "./utils/logger";
 
 const router = Router();
