+++ server/server.js
@@ -1,8 +1,8 @@
 import http from "http";
 
 import app from "./app";
-import config from "./config";
-import logger from "./logger";
+import config from "./utils/config";
+import logger from "./utils/logger";
 
 const server = http.createServer(app);
